SVD analysis with Rank Approximation
Much of the mathematics was provided by Dr. ILIE UGARCOVICI from DePaul University. This is a SVD analyzing tool that can use a variety of different terms or self generated terms from text overall. Essentially, this takes text and performs a unique count to create a matrix of most prominent words and uses eigenvectors to understand the constituent elements of the matrix. The elements are the left singular values (U), eigenvectors (S), and the right singulars (V^t being  transposed to work with the rows of the matrix). Based on a ranked approximation, one can retrieve the eigenvectors  of the given text and use them as vectors within a 2D coordinated space. Then using the query coordinates through a similiar process, one can get the vector coordinates of the query words. Then using  a cosine similarity  ( (dot product of the query coords. and document coords) / multiplication of the norms of both items), one can determine the closest match of the query to the document by the numerical result.